# src/main/openapi/openapispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/aggregations/total-per-person": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "totalsPerPersonByYear_1",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AggregationPerson"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/aggregations/total-per-month": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "totalsPerMonthByYear",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AggregationMonth"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/aggregations/total-per-client": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "revenuePerClientByYear",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AggregationClient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/aggregations/person-nonproductive-hours-per-day": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "personNonProductiveHoursPerDay",
        "parameters": [
          {
            "name": "personId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NonProductiveHours"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/aggregations/leave-day-report": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "leaveDayReportByYear",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AggregationLeaveDay"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/aggregations/leave-day-report-me": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "leaveDayReportMeByYear",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationLeaveDay"
                }
              }
            }
          }
        }
      }
    },
    "/api/aggregations/holiday-details-me": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "holidayDetailsMeYear",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonHolidayDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/aggregations/client-hour-overview": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "hourClientOverviewEmployee",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AggregationClientPersonOverview"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/aggregations/client-assignment-hour-overview": {
      "get": {
        "tags": ["aggregation-controller"],
        "operationId": "hourAssignmentClientOverviewEmployee",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AggregationClientPersonAssignmentOverview"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AggregationIdentifier": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AggregationPerson": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "contractTypes": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sickDays": {
            "type": "number"
          },
          "workDays": {
            "type": "number"
          },
          "assignment": {
            "type": "integer",
            "format": "int32"
          },
          "event": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "leaveDayUsed": {
            "type": "number"
          },
          "leaveDayBalance": {
            "type": "number"
          },
          "paidParentalLeaveUsed": {
            "type": "number"
          },
          "unpaidParentalLeaveUsed": {
            "type": "number"
          },
          "revenue": {
            "$ref": "#/components/schemas/AggregationPersonClientRevenueOverview"
          },
          "cost": {
            "type": "number"
          }
        }
      },
      "AggregationPersonClientRevenueItem": {
        "type": "object",
        "properties": {
          "client": {
            "$ref": "#/components/schemas/AggregationIdentifier"
          },
          "revenue": {
            "type": "number"
          }
        }
      },
      "AggregationPersonClientRevenueOverview": {
        "type": "object",
        "properties": {
          "clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AggregationPersonClientRevenueItem"
            }
          },
          "total": {
            "type": "number"
          }
        }
      },
      "AggregationMonth": {
        "type": "object",
        "properties": {
          "yearMonth": {
            "type": "string"
          },
          "countContractInternal": {
            "type": "integer",
            "format": "int32"
          },
          "countContractManagement": {
            "type": "integer",
            "format": "int32"
          },
          "countContractExternal": {
            "type": "integer",
            "format": "int32"
          },
          "forecastRevenueGross": {
            "type": "number"
          },
          "forecastRevenueNet": {
            "type": "number"
          },
          "forecastHoursGross": {
            "type": "number"
          },
          "actualRevenue": {
            "type": "number"
          },
          "actualHours": {
            "type": "number"
          },
          "actualCostContractInternal": {
            "type": "number"
          },
          "actualCostContractExternal": {
            "type": "number"
          },
          "actualCostContractManagement": {
            "type": "number"
          },
          "actualCostContractService": {
            "type": "number"
          },
          "actualRevenueInternal": {
            "type": "number"
          },
          "actualRevenueExternal": {
            "type": "number"
          },
          "actualRevenueManagement": {
            "type": "number"
          }
        }
      },
      "AggregationClient": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "revenueGross": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "NonProductiveHours": {
        "type": "object",
        "properties": {
          "sickHours": {
            "type": "number",
            "format": "double"
          },
          "holidayHours": {
            "type": "number",
            "format": "double"
          },
          "paidParentalLeaveHours": {
            "type": "number",
            "format": "double"
          },
          "unpaidParentalLeaveHours": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AggregationLeaveDay": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "contractHours": {
            "type": "number",
            "format": "double"
          },
          "plusHours": {
            "type": "number",
            "format": "double"
          },
          "holidayHours": {
            "type": "number",
            "format": "double"
          },
          "paidParentalLeaveHours": {
            "type": "number",
            "format": "double"
          },
          "unpaidParentalLeaveHours": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PersonHolidayDetails": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "holidayHoursFromContract": {
            "type": "number",
            "format": "double"
          },
          "plusHours": {
            "type": "number",
            "format": "double"
          },
          "holidayHoursDone": {
            "type": "number",
            "format": "double"
          },
          "holidayHoursApproved": {
            "type": "number",
            "format": "double"
          },
          "holidayHoursRequested": {
            "type": "number",
            "format": "double"
          },
          "totalHoursAvailable": {
            "type": "number",
            "format": "double"
          },
          "totalHoursUsed": {
            "type": "number",
            "format": "double"
          },
          "totalHoursRemaining": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AggregationClientPersonItem": {
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/AggregationIdentifier"
          },
          "hours": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AggregationClientPersonOverview": {
        "type": "object",
        "properties": {
          "client": {
            "$ref": "#/components/schemas/AggregationIdentifier"
          },
          "aggregationPerson": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AggregationClientPersonItem"
            }
          },
          "totals": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        }
      },
      "AggregationClientPersonAssignmentItem": {
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/AggregationIdentifier"
          },
          "assignment": {
            "$ref": "#/components/schemas/AggregationIdentifier"
          },
          "hours": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AggregationClientPersonAssignmentOverview": {
        "type": "object",
        "properties": {
          "client": {
            "$ref": "#/components/schemas/AggregationIdentifier"
          },
          "aggregationPersonAssignment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AggregationClientPersonAssignmentItem"
            }
          },
          "totals": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        }
      },
...
    }
  }
}